build-i686: do not try to build packages that were dropped
As was the case with the `git flow` package, when a `PKGBUILD` file is
dropped, a naive `git diff` invocation would suggest to try to build it.

Such a naive invocation is in the `build-i686` definition, courtesy of
yours truly.

Let's restrict the `diff` output via a diff filter that excludes deleted
files. Sadly, this is only possible by listing _all other_ values of the
[`--diff-filter`
option](https://git-scm.com/docs/git-diff#Documentation/git-diff.txt---diff-filterACDMRTUXB).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>