pcre2: skip that pesky `fr_FR` test altogether
There is a test in PCRE2 that has given us grief for long enough when
building the MSYS variant of PCRE2: The ISO-8859-1 one. The problem is a
difference in assumptions between PCRE2 and the Cygwin runtime (or for
that matter, the MSYS2 runtime).

For years, we tried to work around it, but it is finally time to admit
that keeping the workaround working is not worth the effort. All it
would do would be to continually document the difference in opinion, and
it would not add the confidence a test should add.

So let's drop the workaround, and skip the ISO-8859-1 test at long last.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>